-Modify query to handle both fields -Display all of them together with uniqBy since there are a lot of duplicate
